Move ClusterNodeFileN out of config package
