+++ src/test/java/com/hivemq/extensions/cluster/discovery/s3/S3DiscoveryCallbackTest.java
@@ -25,7 +25,6 @@ import com.hivemq.extension.sdk.api.services.cluster.parameter.ClusterDiscoveryO
 import com.hivemq.extension.sdk.api.services.cluster.parameter.ClusterNodeAddress;
 import com.hivemq.extensions.cluster.discovery.s3.aws.HiveMQS3Client;
 import com.hivemq.extensions.cluster.discovery.s3.aws.S3BucketResponse;
-import com.hivemq.extensions.cluster.discovery.s3.config.ClusterNodeFile;
 import com.hivemq.extensions.cluster.discovery.s3.config.ConfigurationReader;
 import com.hivemq.extensions.cluster.discovery.s3.config.S3Config;
 import com.hivemq.extensions.cluster.discovery.s3.util.ClusterNodeFileUtil;
